# Conflicts:
#	core-bundle/config/listener.yaml
#	core-bundle/contao/templates/twig/content_element/description_list.html.twig

+++ core-bundle/contao/forms/Form.php
@@ -38,6 +38,7 @@ use Symfony\Component\HttpFoundation\Response;
 class Form extends Hybrid
 {
 	public const SESSION_KEY = 'contao.form.data';
+
 	public const SESSION_CONFIRMATION_KEY = 'contao.form.confirmation';
 
 	/**
@@ -187,6 +188,8 @@ class Form extends Hybrid
 					$arrFields[] = $objFields->current();
 				}
 			}
+
+			System::getContainer()->get('contao.cache.entity_tags')->tagWith($objFields);
 		}
 
 		// HOOK: compile form fields
@@ -194,8 +197,7 @@ class Form extends Hybrid
 		{
 			foreach ($GLOBALS['TL_HOOKS']['compileFormFields'] as $callback)
 			{
-				$this->import($callback[0]);
-				$arrFields = $this->{$callback[0]}->{$callback[1]}($arrFields, $formId, $this);
+				$arrFields = System::importStatic($callback[0])->{$callback[1]}($arrFields, $formId, $this);
 			}
 		}
 
@@ -239,8 +241,7 @@ class Form extends Hybrid
 				{
 					foreach ($GLOBALS['TL_HOOKS']['loadFormField'] as $callback)
 					{
-						$this->import($callback[0]);
-						$objWidget = $this->{$callback[0]}->{$callback[1]}($objWidget, $formId, $this->arrData, $this);
+						$objWidget = System::importStatic($callback[0])->{$callback[1]}($objWidget, $formId, $this->arrData, $this);
 					}
 				}
 
@@ -254,8 +255,7 @@ class Form extends Hybrid
 					{
 						foreach ($GLOBALS['TL_HOOKS']['validateFormField'] as $callback)
 						{
-							$this->import($callback[0]);
-							$objWidget = $this->{$callback[0]}->{$callback[1]}($objWidget, $formId, $this->arrData, $this);
+							$objWidget = System::importStatic($callback[0])->{$callback[1]}($objWidget, $formId, $this->arrData, $this);
 						}
 					}
 
@@ -311,7 +311,7 @@ class Form extends Hybrid
 					$file->delete();
 				}
 
-				if (is_file($upload['tmp_name']))
+				if (isset($upload['tmp_name']) && is_file($upload['tmp_name']))
 				{
 					unlink($upload['tmp_name']);
 				}
@@ -390,8 +390,7 @@ class Form extends Hybrid
 		{
 			foreach ($GLOBALS['TL_HOOKS']['prepareFormData'] as $callback)
 			{
-				$this->import($callback[0]);
-				$this->{$callback[0]}->{$callback[1]}($arrSubmitted, $arrLabels, $arrFields, $this, $arrFiles);
+				System::importStatic($callback[0])->{$callback[1]}($arrSubmitted, $arrLabels, $arrFields, $this, $arrFiles);
 			}
 		}
 
@@ -439,7 +438,7 @@ class Form extends Hybrid
 				if ($this->format == 'csv' || $this->format == 'csv_excel')
 				{
 					$keys[] = $k;
-					$values[] = (\is_array($v) ? implode(',', $v) : $v);
+					$values[] = \is_array($v) ? implode(',', $v) : $v;
 				}
 			}
 
@@ -497,6 +496,9 @@ class Form extends Hybrid
 			// Attach XML file
 			if ($this->format == 'xml')
 			{
+				// Encode the values (see #6053)
+				array_walk_recursive($fields, static function (&$value) { $value = htmlspecialchars($value, ENT_QUOTES|ENT_SUBSTITUTE|ENT_XML1); });
+
 				$objTemplate = new FrontendTemplate('form_xml');
 				$objTemplate->fields = $fields;
 				$objTemplate->charset = System::getContainer()->getParameter('kernel.charset');
@@ -548,10 +550,11 @@ class Form extends Hybrid
 		// Store the values in the database
 		if ($this->storeValues && $this->targetTable)
 		{
+			$db = Database::getInstance();
 			$arrSet = array();
 
 			// Add the timestamp
-			if ($this->Database->fieldExists('tstamp', $this->targetTable))
+			if ($db->fieldExists('tstamp', $this->targetTable))
 			{
 				$arrSet['tstamp'] = time();
 			}
@@ -589,8 +592,7 @@ class Form extends Hybrid
 			{
 				foreach ($GLOBALS['TL_HOOKS']['storeFormData'] as $callback)
 				{
-					$this->import($callback[0]);
-					$arrSet = $this->{$callback[0]}->{$callback[1]}($arrSet, $this);
+					$arrSet = System::importStatic($callback[0])->{$callback[1]}($arrSet, $this);
 				}
 			}
 
@@ -607,7 +609,7 @@ class Form extends Hybrid
 			}
 
 			// Do not use Models here (backwards compatibility)
-			$this->Database->prepare("INSERT INTO " . $this->targetTable . " %s")->set($arrSet)->execute();
+			$db->prepare("INSERT INTO " . $this->targetTable . " %s")->set($arrSet)->execute();
 		}
 
 		// HOOK: process form data callback
@@ -615,17 +617,14 @@ class Form extends Hybrid
 		{
 			foreach ($GLOBALS['TL_HOOKS']['processFormData'] as $callback)
 			{
-				$this->import($callback[0]);
-				$this->{$callback[0]}->{$callback[1]}($arrSubmitted, $this->arrData, $arrFiles, $arrLabels, $this);
+				System::importStatic($callback[0])->{$callback[1]}($arrSubmitted, $this->arrData, $arrFiles, $arrLabels, $this);
 			}
 		}
 
 		// Add a log entry
 		if (System::getContainer()->get('contao.security.token_checker')->hasFrontendUser())
 		{
-			$this->import(FrontendUser::class, 'User');
-
-			System::getContainer()->get('monolog.logger.contao.forms')->info('Form "' . $this->title . '" has been submitted by "' . $this->User->username . '".');
+			System::getContainer()->get('monolog.logger.contao.forms')->info('Form "' . $this->title . '" has been submitted by "' . FrontendUser::getInstance()->username . '".');
 		}
 		else
 		{
